+++ playground/lib/pages/image.dart
@@ -4,6 +4,7 @@ import 'package:shadcn_ui/shadcn_ui.dart';
 enum ImageStyle {
   local,
   remote,
+  remoteWithHeaders,
   svg,
 }
 
@@ -29,6 +30,13 @@ class ImagePage extends StatelessWidget {
             ImageStyle.remote => const ShadImage.square(
                 'https://avatars.githubusercontent.com/u/124599?v=4',
                 size: 50,
+              ),
+            ImageStyle.remoteWithHeaders => const ShadImage.square(
+                'https://avatars.githubusercontent.com/u/124599?v=4',
+                size: 50,
+                headers: {
+                  'User-Agent': 'Mozilla/5.0',
+                },
               ),
             ImageStyle.svg => const ShadImage.square(
                 'assets/flutter.svg',